Omitted deprecated Projects. Updated menu-driven interface; items 1, 2 and X are functional. Created Generator Helper and XmlFileHelper, both functional.
