+++ FileManagementHelper/DefaultLingoWords.xml
@@ -1,0 +1,136 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<!--Init LingoWords.xml with cSharp lingo-->
+<Words>
+  <Item>
+    <Word>method</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>inheritance</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>class</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>object</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>instance</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>property</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>field</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>constructor</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>dot net</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>array</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>main</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>curly brace</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>string</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>boolean</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>semicolon</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>parse</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>try catch</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>partial</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>return</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>call</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>override</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>keyword</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>get or set</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>static</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>interface</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>base</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>virtual</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>object</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>abstract</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>tryparse</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>encapsulation</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>ADO.NET</Word>
+    <Category>CSharp</Category>
+  </Item>
+  <Item>
+    <Word>Entity Framework</Word>
+    <Category>CSharp</Category>
+  </Item>
+</Words>
