+++ FileManagementHelper/FileManagementHelper.csproj
@@ -43,6 +43,16 @@
   <ItemGroup>
     <Compile Include="DbManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="XmlFileHelper.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="DefaultLingoWords.xml" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\LingoBingoGenerator\LingoBingoGenerator.csproj">
+      <Project>{E671CFB7-630F-404D-8B0E-A94D8BE4F02D}</Project>
+      <Name>LingoBingoGenerator</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
