+++ UnitTester/LingoGeneratorTests.cs
@@ -3,8 +3,6 @@ using LingoBingoGenerator;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace LingoBingoGenerator.Tests
 {
